scroll-animations/scroll-timeline-phases.tentative.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Test basic functionality of scroll timeline phases.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/web-animations/testcommon.js"></script>
<script src="testcommon.js"></script>
<style>
  .scroller {
    overflow: auto;
    height: 100px;
    width: 100px;
  }
  .contents {
    height: 1000px;
    width: 100%;
  }
</style>
<div id="log"></div>
<script>
  'use strict';

  promise_test(async t => {
    const timeline = createScrollTimeline(t);
    assert_equals(timeline.phase, "inactive");
  }, 'Scroll timeline starts in "inactive" phase.');

  promise_test(async t => {
    const timeline = createScrollTimeline(t);
    assert_throws_js(TypeError, () => {
      timeline.phase = "after";
    });
  }, 'Setting scroll timeline phase (which is readonly) throws TypeError.');

  const test_cases = {
    before_start: {
      name: "before start",
      scroll_percent: 0.1,
      timeline_phase: "before"
    },
    at_start: {
      name: "at start",
      scroll_percent: 0.2,
      timeline_phase: "active"
    },
    in_range: {
      name: "in range",
      scroll_percent: 0.5,
      timeline_phase: "active"
    },
    at_end: {
      name: "at end",
      scroll_percent: 0.8,
      timeline_phase: "after"
    },
    after_end: {
      name: "after end",
      scroll_percent: 0.9,
      timeline_phase: "after"
    }
  }

  for (const test_case_key in test_cases){
    const test_case = test_cases[test_case_key];
    promise_test(async t => {
      const timeline = createScrollTimelineWithOffsets(t);
      const scroller = timeline.scrollSource;
      const maxScroll = scroller.scrollHeight - scroller.clientHeight;

      scroller.scrollTop = test_case.scroll_percent * maxScroll;

      assert_equals(
        timeline.phase,
        test_case.timeline_phase,
        "timeline.phase"
      );
    }, "Timeline phase while scroll offset is " + test_case.name);
  }
</script>
